fix: resolve 77 CodeQL code scanning alerts
iccanalyzer-lite (2 fixes):
- IccHeuristicsRawPost.cpp: restructure LUT tag sig index to prevent
  unsigned subtraction before bounds check (4 buffer-overflow alerts)
- IccImageAnalyzer.cpp: use fstat(fileno(fp)) instead of stat()+fopen()
  to eliminate TOCTOU race condition (1 alert)

CFL fuzzers (6 fixes):
- fuzz_utils.h: cast data bytes to uint32_t before left-shift to prevent
  signed-shift-overflow UB (4 alerts)
- icc_toxml_fuzzer.cpp, icc_dump_fuzzer.cpp: use open(O_CREAT, 0600)
  instead of fopen("wb") to prevent world-writable temp files (2 alerts)
- icc_fromcube_fuzzer.cpp: copy interior pointer values to local arrays
  to fix use-after-expired-lifetime (9 alerts), use memcmp for float
  equality matching upstream exact-bit semantics (8 alerts)
- icc_specsep_fuzzer.cpp: use memcmp for TIFF resolution float comparison
  matching upstream exact semantics (2 alerts)
- icc_applynamedcmm_fuzzer.cpp: remove unused pcs and SrcNameBuf
  variables (2 alerts)

MCP server (2 fixes):
- web_ui.py: add resolve()+relative_to() defense-in-depth for path
  traversal prevention in /api/list-xml (1 alert)
- icc_profile_mcp.py: add explanatory comments to except-pass blocks
  for process cleanup and TTL file removal (4 alerts)

Dismissed (42 alerts):
- 32 colorbleed_tools alerts (intentionally unsafe by design)
- 5 upstream iccDEV alerts (not our code)
- 3 CFL XXE alerts (fuzzers intentionally process XML)
- 2 informational alerts (custom query, style)

+++ cfl/fuzz_utils.h
@@ -34,21 +34,21 @@ static inline size_t fuzz_build_path(char *buf, size_t bufsize,
 // Returns true if profile passes validation, false to reject.
 static inline bool fuzz_validate_icc_tags(const uint8_t *data, size_t size) {
     if (size < 132) return false;
-    uint32_t tagCount = (data[128] << 24) | (data[129] << 16) |
-                        (data[130] << 8) | data[131];
+    uint32_t tagCount = ((uint32_t)data[128] << 24) | ((uint32_t)data[129] << 16) |
+                        ((uint32_t)data[130] << 8) | (uint32_t)data[131];
     if (tagCount > 200) return false;
     for (uint32_t t = 0; t < tagCount; t++) {
         size_t base = 132 + t * 12;
         if (base + 12 > size) return false;
-        uint32_t tOff  = (data[base+4] << 24) | (data[base+5] << 16) |
-                         (data[base+6] << 8) | data[base+7];
-        uint32_t tSize = (data[base+8] << 24) | (data[base+9] << 16) |
-                         (data[base+10] << 8) | data[base+11];
+        uint32_t tOff  = ((uint32_t)data[base+4] << 24) | ((uint32_t)data[base+5] << 16) |
+                         ((uint32_t)data[base+6] << 8) | (uint32_t)data[base+7];
+        uint32_t tSize = ((uint32_t)data[base+8] << 24) | ((uint32_t)data[base+9] << 16) |
+                         ((uint32_t)data[base+10] << 8) | (uint32_t)data[base+11];
         if (tOff > size || tSize > size) return false;
         if (tOff + tSize < tOff) return false;  // overflow
     }
-    uint32_t hdrSize = (data[0] << 24) | (data[1] << 16) |
-                       (data[2] << 8) | data[3];
+    uint32_t hdrSize = ((uint32_t)data[0] << 24) | ((uint32_t)data[1] << 16) |
+                       ((uint32_t)data[2] << 8) | (uint32_t)data[3];
     if (hdrSize > size) return false;
     return true;
 }
